// src/app/core/data/drawings.generated.ts
// Generated automatically. Do not edit directly.
export const localDrawings = [
  {
    "title": "A (6)",
    "category": "Color Art",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/ColorArt/2026-09-14/a (6).jpeg",
    "alt": "A (6)"
  },
  {
    "title": "A (5)",
    "category": "Color Art",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/ColorArt/2026-09-14/a (5).jpeg",
    "alt": "A (5)"
  },
  {
    "title": "A (4)",
    "category": "Color Art",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/ColorArt/2026-09-14/a (4).jpeg",
    "alt": "A (4)"
  },
  {
    "title": "A (3)",
    "category": "Color Art",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/ColorArt/2026-09-14/a (3).jpeg",
    "alt": "A (3)"
  },
  {
    "title": "A (2)",
    "category": "Color Art",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/ColorArt/2026-09-14/a (2).jpeg",
    "alt": "A (2)"
  },
  {
    "title": "A (1)",
    "category": "Color Art",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/ColorArt/2026-09-14/a (1).jpeg",
    "alt": "A (1)"
  },
  {
    "title": "Images",
    "category": "Pencil Sketches",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/PencilSketches/2026-09-14/images.jpg",
    "alt": "Images"
  },
  {
    "title": "A (3)",
    "category": "Pencil Sketches",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/PencilSketches/2026-09-14/a (3).jpeg",
    "alt": "A (3)"
  },
  {
    "title": "A (2)",
    "category": "Pencil Sketches",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/PencilSketches/2026-09-14/a (2).jpeg",
    "alt": "A (2)"
  },
  {
    "title": "A (1)",
    "category": "Pencil Sketches",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/PencilSketches/2026-09-14/a (1).jpeg",
    "alt": "A (1)"
  },
  {
    "title": "A (1)",
    "category": "Pencil Sketches",
    "year": "2026",
    "date": "2026-09-14",
    "image": "/drawings/PencilSketches/2026-09-14/a (1).avif",
    "alt": "A (1)"
  }
] as const;
export const featuredDrawing = {
  "title": "Feature",
  "category": "Featured Drawing",
  "year": "",
  "date": "",
  "image": "/drawings/FeatureDrawing/feature.jpeg",
  "alt": "Feature"
} as const;